Merging Passport guards

+++ app/Http/Controllers/BarberController.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\StoreBarberRequest;
 use App\Models\Barber;
 use App\Models\Barbershop;
 use Illuminate\Http\Request;
@@ -11,24 +10,82 @@ use Illuminate\Support\Facades\File;
 class BarberController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Store a new barber.
+     *
+     * @OA\Post(
+     *     path="/api/barbers",
+     *     summary="Store a new barber.",
+     *     description="Store a new barber with name, barbershop id, and image.",
+     *     tags={"Barber"},
+     *
+     *     @OA\RequestBody(
+     *         required=true,
+     *
+     *         @OA\MediaType(
+     *             mediaType="multipart/form-data",
+     *
+     *             @OA\Schema(
+     *                 required={"name", "barbershop_id", "image"},
+     *
+     *                 @OA\Property(
+     *                     property="name",
+     *                     type="string",
+     *                     description="The name of the barber."
+     *                 ),
+     *                 @OA\Property(
+     *                     property="barbershop_id",
+     *                     type="integer",
+     *                     description="The ID of the barbershop that the barber belongs to."
+     *                 ),
+     *                 @OA\Property(
+     *                     property="image",
+     *                     type="file",
+     *                     description="The image of the barber."
+     *                 )
+     *             )
+     *         )
+     *     ),
+     *
+     *     @OA\Response(
+     *         response=201,
+     *         description="Barber created successfully",
+     *
+     *         @OA\JsonContent(
+     *
+     *             @OA\Property(
+     *                 property="message",
+     *                 type="string",
+     *                 example="Barber created successfully"
+     *             ),
+     *             @OA\Property(
+     *                 property="barber"
+     *             )
+     *         )
+     *     ),
+     *
+     *     @OA\Response(
+     *         response=422,
+     *         description="Unprocessable Entity",
+     *
+     *         @OA\JsonContent(
+     *
+     *             @OA\Property(
+     *                 property="message",
+     *                 type="string",
+     *                 example="The given data was invalid."
+     *             ),
+     *             @OA\Property(
+     *                 property="errors",
+     *                 type="object"
+     *             )
+     *         )
+     *     ),
+     *     security={
+     *         {"bearerAuth": {}}
+     *     }
+     * )
      */
-    public function index()
-    {
-    }
-
-    //get all barbers of a barbershop
-    public function getBarbersOfBarbershop($barbershop_id)
-    {
-        $barberShop = BarberShop::find($barbershop_id);
-
-        return $barberShop->barbers;
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(StoreBarberRequest $request)
+    public function store(Request $request)
     {
         $this->validate(
             $request,
@@ -56,16 +113,115 @@ class BarberController extends Controller
     }
 
     /**
-     * Update the specified resource in storage.
+     * Update an existing barber.
+     *
+     * @OA\Put(
+     *     path="/api/barbers/{barber_id}",
+     *     summary="Update an existing barber.",
+     *     description="Update an existing barber with name, barbershop id, and image.",
+     *     tags={"Barber"},
+     *
+     *     @OA\Parameter(
+     *         name="barber_id",
+     *         in="path",
+     *         required=true,
+     *         description="The ID of the barber to update.",
+     *
+     *         @OA\Schema(
+     *             type="integer",
+     *             format="int64",
+     *             example=1
+     *         )
+     *     ),
+     *
+     *     @OA\RequestBody(
+     *         required=true,
+     *
+     *         @OA\MediaType(
+     *             mediaType="multipart/form-data",
+     *
+     *             @OA\Schema(
+     *                 required={"name", "barbershop_id"},
+     *
+     *                 @OA\Property(
+     *                     property="name",
+     *                     type="string",
+     *                     description="The name of the barber."
+     *                 ),
+     *                 @OA\Property(
+     *                     property="barbershop_id",
+     *                     type="integer",
+     *                     description="The ID of the barbershop that the barber belongs to."
+     *                 ),
+     *                 @OA\Property(
+     *                     property="image",
+     *                     type="file",
+     *                     description="The image of the barber."
+     *                 )
+     *             )
+     *         )
+     *     ),
+     *
+     *     @OA\Response(
+     *         response=200,
+     *         description="Barber updated successfully",
+     *
+     *         @OA\JsonContent(
+     *
+     *             @OA\Property(
+     *                 property="message",
+     *                 type="string",
+     *                 example="Barber updated successfully"
+     *             )
+     *         )
+     *     ),
+     *
+     *     @OA\Response(
+     *         response=422,
+     *         description="Unprocessable Entity",
+     *
+     *         @OA\JsonContent(
+     *
+     *             @OA\Property(
+     *                 property="message",
+     *                 type="string",
+     *                 example="The given data was invalid."
+     *             ),
+     *             @OA\Property(
+     *                 property="errors",
+     *                 type="object"
+     *             )
+     *         )
+     *     ),
+     *
+     *     @OA\Response(
+     *         response=404,
+     *         description="Barber not found",
+     *
+     *         @OA\JsonContent(
+     *
+     *             @OA\Property(
+     *                 property="message",
+     *                 type="string",
+     *                 example="Barber not found"
+     *             )
+     *         )
+     *     ),
+     *     security={
+     *         {"bearerAuth": {}}
+     *     }
+     * )
      */
-    public function update($request, $barber_id)
+    public function update(Request $request, $barber_id)
     {
-        $this->validate($request,
+        $this->validate(
+            $request,
             [
                 'name' => 'required',
                 'barbershop_id' => 'required',
                 'image' => 'nullable|image|mimes:jpeg,png,jpg,gif,svg',
-            ]);
+            ]
+        );
         $barber = Barber::find($barber_id);
         $barber->name = $request->name;
         $barber->barbershop_id = $request->barbershop_id;
@@ -84,18 +240,86 @@ class BarberController extends Controller
         $barber->save();
 
         return response()->json(['message' => 'Barber updated successfully']);
-
-
-
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Delete a barber.
+     *
+     * @OA\Delete(
+     *     path="/api/barbers/{barber_id}",
+     *     summary="Delete a barber.",
+     *     description="Delete a barber by ID.",
+     *     tags={"Barber"},
+     *
+     *     @OA\Parameter(
+     *         name="barber_id",
+     *         in="path",
+     *         required=true,
+     *         description="ID of the barber to delete",
+     *
+     *         @OA\Schema(
+     *             type="integer",
+     *             format="int64"
+     *         )
+     *     ),
+     *
+     *     @OA\Response(
+     *         response=200,
+     *         description="Barber deleted successfully.",
+     *
+     *         @OA\JsonContent(
+     *
+     *             @OA\Property(
+     *                 property="status",
+     *                 type="integer",
+     *                 example=200
+     *             ),
+     *             @OA\Property(
+     *                 property="message",
+     *                 type="string",
+     *                 example="The barber has been deleted!"
+     *             )
+     *         )
+     *     ),
+     *
+     *     @OA\Response(
+     *         response=404,
+     *         description="Barber not found.",
+     *
+     *         @OA\JsonContent(
+     *
+     *             @OA\Property(
+     *                 property="status",
+     *                 type="integer",
+     *                 example=404
+     *             ),
+     *             @OA\Property(
+     *                 property="errors",
+     *                 type="string",
+     *                 example="No barber found to be deleted!"
+     *             )
+     *         )
+     *     ),
+     *     security={
+     *         {"bearerAuth": {}}
+     *     }
+     * )
      */
-    public function destroy(Barber $barber)
+    public function destroy($barber_id)
     {
-        $barber->delete();
+        $barber = Barber::find($barber_id);
+        if (is_null($barber) || empty($barber)) {
+            return response()->json([
+                'status' => 404,
+                'errors' => 'No barber found to be deleted!',
+            ]);
+        } else {
+            $barber->delete();
 
-        return response()->json(['message' => 'Barber deleted successfully']);
+            return response()->json([
+                'status' => 200,
+                'message' => 'The barber has been deleted!',
+            ]);
+        }
     }
 }
